Removed hardcoded input layout & shader loading. Support for loading & setting shaders and input layouts through messages added. Updated loading graphics resources to use messages instead of direct function calls. Removed redundant IConstantBuffer and IVertex classes

+++ src/Cork/Cork/BinaryData.cpp
@@ -1,6 +1,6 @@
 #include "BinaryData.h"
 
-BinaryData::BinaryData(int sizeInBytes) : sizeInBytes(sizeInBytes), buffer(new char[sizeInBytes]), currentPosition(buffer), currentBufferSize(0) {}
+BinaryData::BinaryData(size_t sizeInBytes) : sizeInBytes(sizeInBytes), buffer(new char[sizeInBytes]), currentPosition(buffer), currentBufferSize(0) {}
 
 BinaryData::~BinaryData() {
   delete[] buffer;
